Ouroboros.Network.Handshake.Acceptable

Documentation

data Accept vData

A Maybe like type which better explains its purpose.

Constructors

Accept vData

Refuse !Text

class Acceptable v where

Methods

acceptableVersion :: v -> v -> Accept v

The acceptableVersion function ought to be symmetric, this guarantees that local and remote sides will agree on the same data.