Ouroboros.Network.NodeToClient.Version

Documentation

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_9
enabled CardanoNodeToClientVersion7, i.e., Alonzo

NodeToClientV_10
added GetChainBlockNo and GetChainPoint queries

NodeToClientV_11
added GetRewardInfoPools Block query

NodeToClientV_12
added LocalTxMonitor mini-protocol

NodeToClientV_13
enabled CardanoNodeToClientVersion9, i.e., Babbage

NodeToClientV_14
added GetPoolDistr, GetPoolState, GetSnapshots

NodeToClientV_15
added query to NodeToClientVersionData

NodeToClientV_16
add ImmutableTip to LocalStateQuery, enabled CardanoNodeToClientVersion11, i.e., Conway and GetStakeDelegDeposits.

data NodeToClientVersionData

Version data for NodeToClient protocol v1

Constructors

NodeToClientVersionData

Fields

networkMagic :: !NetworkMagic

query :: !Bool

nodeToClientCodecCBORTerm :: NodeToClientVersion -> CodecCBORTerm Text NodeToClientVersionData

nodeToClientVersionCodec :: CodecCBORTerm (Text, Maybe Int) NodeToClientVersion

We set 16ths bit to distinguish NodeToNodeVersion and NodeToClientVersion. This way connecting wrong protocol suite will fail during Handshake negotiation

This is done in backward compatible way, so NodeToClientV_1 encoding is not changed.