Ouroboros.Network.NodeToNode.Version

Documentation

data NodeToNodeVersion

Enumeration of node to node protocol versions.

Constructors

NodeToNodeV_7
Changes:

new KeepAlive codec

Enable CardanoNodeToNodeVersion5, i.e., Alonzo

NodeToNodeV_8
Changes:

Enable block diffusion pipelining in ChainSync and BlockFetch logic.

NodeToNodeV_9
Changes:

Enable CardanoNodeToNodeVersion6, i.e., Babbage

NodeToNodeV_10
Changes:

Enable full duplex connections.

NodeToNodeV_11
Changes:

Adds a new extra parameter to handshake: PeerSharing This version is needed to support the new Peer Sharing miniprotocol older versions that are negotiated will appear as not participating in Peer Sharing to newer versions.

Adds query to NodeToClientVersionData.

NodeToNodeV_12
Changes:

Enable CardanoNodeToNodeVersion7, i.e., Conway

NodeToNodeV_13
Changes:

Removed PeerSharingPrivate constructor

Fixed Codec to disable PeerSharing with buggy versions 11 and 12.

Disable PeerSharing with InitiatorOnly nodes, since they do not run peer sharing server side and can not reply to requests.

data NodeToNodeVersionData

Version data for NodeToNode protocol

Constructors

NodeToNodeVersionData

Fields

networkMagic :: !NetworkMagic

diffusionMode :: !DiffusionMode

peerSharing :: !PeerSharing

query :: !Bool

data DiffusionMode

The flag which indicates whether the node runs only initiator or both initiator or responder node.

This data structure has two proposes:

instruct the diffusion layer if it should listen on incoming connections;

it is communicated via NodeToNodeVersionData during handshake negotiation. In non-p2p mode we always send InitiatorOnlyDiffusionMode, in p2p mode we send exactly what the diffusion is given. In non-p2p mode every connection outbound port is ephemeral, the remote side cannot connect to it, however in p2p mode the outbound port is actually the port on which the node is listening (if it runs in InitiatorAndResponderDiffusionMode).

Constructors

InitiatorOnlyDiffusionMode

InitiatorAndResponderDiffusionMode

data ConnectionMode

Constructors

UnidirectionalMode

DuplexMode

nodeToNodeVersionCodec :: CodecCBORTerm (Text, Maybe Int) NodeToNodeVersion

nodeToNodeCodecCBORTerm :: NodeToNodeVersion -> CodecCBORTerm Text NodeToNodeVersionData

isPipeliningEnabled :: NodeToNodeVersion -> WhetherReceivingTentativeBlocks

Check whether a version enabling diffusion pipelining has been negotiated.

TODO: this ought to be defined in `ouroboros-consensus` or `ouroboros-consensus-diffusion`