Ouroboros.Network.PeerSelection.PeerAdvertise

Documentation

data PeerAdvertise

Should this peer be advertised to other peers asking for known peers? For certain peers specified by configuration it would be an appropriate policy to keep them private.

Constructors

DoNotAdvertisePeer

DoAdvertisePeer