Ouroboros.Network.PeerSelection.PeerSharing

Documentation

data PeerSharing

Is a peer willing to participate in Peer Sharing? If yes are others allowed to share this peer's address? Information about the node comes from the configuration file, while information about other nodes is received via handshake.

NOTE: This information is only useful if P2P flag is enabled.

Constructors

PeerSharingDisabled
Peer does not participate in Peer Sharing at all

PeerSharingEnabled
Peer participates in Peer Sharing