Ouroboros.Network.PeerSelection.RelayAccessPoint

Documentation

data DomainAccessPoint

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAccessPoint

Fields

dapDomain :: !Domain

dapPortNumber :: !PortNumber

data RelayAccessPoint

A relay can have either an IP address and a port number or a domain with a port number

Constructors

RelayAccessDomain !Domain !PortNumber

RelayAccessAddress !IP !PortNumber

Bundled Patterns

pattern RelayDomainAccessPoint :: DomainAccessPoint -> RelayAccessPoint
RelayDomainAccessPoint a bidirectional pattern which links RelayAccessDomain and DomainAccessPoint.

data IP

Constructors

IPv4

Fields

ipv4 :: IPv4

IPv6

Fields

ipv6 :: IPv6

Socket type re-exports

data PortNumber