Ouroboros.Network.Util.ShowProxy

Documentation

class ShowProxy p where

Minimal complete definition

Nothing

Methods

showProxy ∷ Proxy p → String

default showProxy ∷ Typeable p ⇒ Proxy p → String

data Proxy (t ∷ k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy