Ouroboros.Network.PeerSelection.LedgerPeers.Type

Documentation

newtype PoolStake

The relative stake of a stakepool in relation to the total amount staked. A value in the [0, 1] range.

Constructors

PoolStake

Fields

unPoolStake ∷ Rational

newtype AccPoolStake

The accumulated relative stake of a stake pool, like PoolStake but it also includes the relative stake of all preceding pools. A value in the range [0, 1].

Constructors

AccPoolStake

Fields

unAccPoolStake ∷ Rational