Ouroboros.Network.PeerSelection.PeerMetric.Type

Documentation

type HeaderMetricsTracer m = Tracer (STM m) (SlotNo, Time)

Report arrival time of a header for a given slot.

type FetchedMetricsTracer m = Tracer (STM m) (SizeInBytes, SlotNo, Time)

Report block arrival of given size for some slot at a given time.

data ReportPeerMetrics m peerAddr

Constructors

ReportPeerMetrics

Fields

reportHeader ∷ Tracer (STM m) (TraceLabelPeer peerAddr (SlotNo, Time))

reportFetch ∷ Tracer (STM m) (TraceLabelPeer peerAddr (SizeInBytes, SlotNo, Time))

data TraceLabelPeer peerid a

A peer label for use in Tracers. This annotates tracer output as being associated with a given peer identifier.

Constructors

TraceLabelPeer peerid a