Ouroboros.Network.PeerSelection.PeerSharing

Documentation

data PeerSharing

Is a peer willing to participate in Peer Sharing? If yes are others allowed to share this peer's address? Information about the node comes from the configuration file, while information about other nodes is received via handshake.

NOTE: This information is only useful if P2P flag is enabled.

Constructors

NoPeerSharing
Peer does not participate in Peer Sharing at all

PeerSharingPrivate
Peer participates in Peer Sharing but its address should be private

PeerSharingPublic
Peer participates in Peer Sharing

combinePeerInformation :: PeerSharing -> PeerAdvertise -> PeerSharing

encodePortNumber :: PortNumber -> Encoding

decodePortNumber :: Decoder s PortNumber

encodeRemoteAddress :: SockAddr -> Encoding

This encoder should be faithful to the PeerSharing CDDL Specification.

See the network design document for more details

decodeRemoteAddress :: Decoder s SockAddr

This decoder should be faithful to the PeerSharing CDDL Specification.

See the network design document for more details